TokenGraveyard v2.0 - Royalty bypass via HTS allowance staking
Major rewrite introducing royalty-free NFT burial:

Core Changes:
- New TokenStaker module using HTS allowances to bypass royalties
- LazyGasStation integration for centralized $LAZY payments
- Role-based access control (Admin, ContractUser)
- Batch staking support (up to 50 NFTs, batched internally)
- Auto-refill mechanism from LazyGasStation

Contracts:
- TokenGraveyard.sol: Complete rewrite with staking burial
- TokenStaker.sol: New NFT staking module
- LazyGasStation.sol: New payment handler with burn mechanism
- HederaTokenServiceLite.sol: Lightweight HTS wrapper
- Removed legacy HederaTokenService files

Scripts & Tools:
- Interactive deploy.js with post-deployment setup
- CLI tools for info, admin, bury, associate, allowance
- Updated test suite with comprehensive coverage

Documentation:
- CLAUDE.md for AI-assisted development
- UX Implementation Guide for frontend integration
- Updated README with accurate method documentation

+++ contracts/legacy/ExpiryHelper.sol
@@ -2,19 +2,20 @@
 pragma solidity >=0.8.12 <0.9.0;
 pragma experimental ABIEncoderV2;
 
-import "./HederaTokenService.sol";
+import {IHederaTokenService} from "./IHederaTokenService.sol";
 
 abstract contract ExpiryHelper {
-
     function createAutoRenewExpiry(
         address autoRenewAccount,
-        int64 autoRenewPeriod
+        uint32 autoRenewPeriod
     ) internal pure returns (IHederaTokenService.Expiry memory expiry) {
         expiry.autoRenewAccount = autoRenewAccount;
         expiry.autoRenewPeriod = autoRenewPeriod;
     }
 
-    function createSecondExpiry(int64 second) internal pure returns (IHederaTokenService.Expiry memory expiry) {
+    function createSecondExpiry(
+        uint32 second
+    ) internal pure returns (IHederaTokenService.Expiry memory expiry) {
         expiry.second = second;
     }
 }
